libkern update

+++ sys/lib/libkern/muldi3.c
@@ -50,7 +50,7 @@ __RCSID("$NetBSD: muldi3.c,v 1.8 2003/08/07 16:32:09 agc Exp $");
  * Our algorithm is based on the following.  Split incoming quad values
  * u and v (where u,v >= 0) into
  *
- *	u = 2^n u1  *  u0	(n = number of bits in `u_int', usu. 32)
+ *	u = 2^n u1  *  u0	(n = number of bits in `u_long', usu. 32)
  *
  * and 
  *
@@ -81,9 +81,9 @@ __RCSID("$NetBSD: muldi3.c,v 1.8 2003/08/07 16:32:09 agc Exp $");
  *
  * This algorithm is from Knuth vol. 2 (2nd ed), section 4.3.3, p. 278.
  *
- * Since C does not give us a `int * int = quad' operator, we split
- * our input quads into two ints, then split the two ints into two
- * shorts.  We can then calculate `short * short = int' in native
+ * Since C does not give us a `long * long = quad' operator, we split
+ * our input quads into two longs, then split the two ints into two
+ * shorts.  We can then calculate `short * short = longlong' in native
  * arithmetic.
  *
  * Our product should, strictly speaking, be a `long quad', with 128
@@ -101,14 +101,14 @@ __RCSID("$NetBSD: muldi3.c,v 1.8 2003/08/07 16:32:09 agc Exp $");
  * of 2^n in either one will also vanish.  Only `low' need be computed
  * mod 2^2n, and only because of the final term above.
  */
-static quad_t __lmulq(u_int, u_int);
+static quad_t __lmulq(u_long, u_long);
 
 quad_t
 __muldi3(a, b)
 	quad_t a, b;
 {
 	union uu u, v, low, prod;
-	u_int high, mid, udiff, vdiff;
+	u_long high, mid, udiff, vdiff;
 	int negall, negmid;
 #define	u1	u.ul[H]
 #define	u0	u.ul[L]
@@ -141,7 +141,7 @@ __muldi3(a, b)
 		 * Compute the three intermediate products, remembering
 		 * whether the middle term is negative.  We can discard
 		 * any upper bits in high and mid, so we can use native
-		 * u_int * u_int => u_int arithmetic.
+		 * u_long * u_long => u_long arithmetic.
 		 */
 		low.q = __lmulq(u0, v0);
 
@@ -178,21 +178,21 @@ __muldi3(a, b)
  * typically N==16).
  *
  * We use the same algorithm from Knuth, but this time the modulo refinement
- * does not apply.  On the other hand, since N is half the size of an int,
+ * does not apply.  On the other hand, since N is half the size of an long,
  * we can get away with native multiplication---none of our input terms
- * exceeds (UINT_MAX >> 1).
+ * exceeds (ULONG_MAX >> 1).
  *
- * Note that, for u_int l, the quad-precision result
+ * Note that, for u_long l, the quad-precision result
  *
  *	l << N
  *
- * splits into high and low ints as HHALF(l) and LHUP(l) respectively.
+ * splits into high and low longs as HHALF(l) and LHUP(l) respectively.
  */
 static quad_t
-__lmulq(u_int u, u_int v)
+__lmulq(u_long u, u_long v)
 {
-	u_int u1, u0, v1, v0, udiff, vdiff, high, mid, low;
-	u_int prodh, prodl, was;
+	u_long u1, u0, v1, v0, udiff, vdiff, high, mid, low;
+	u_long prodh, prodl, was;
 	union uu prod;
 	int neg;
